requirements.txt
bertopic>=0.17.0
sentence-transformers>=2.7.0
umap-learn>=0.5.6
hdbscan>=0.8.38
pandas>=2.0.0
scikit-learn>=1.4.0
torch>=2.0.0
tqdm>=4.65.0
matplotlib>=3.8.0
pyyaml>=6.0
ipykernel
nbformat